+++ patches/src/main/kotlin/app/morphe/patches/instagram/reels/Fingerprint.kt
@@ -2,11 +2,11 @@ package app.morphe.patches.instagram.reels
 
 import app.morphe.patcher.Fingerprint
 
-internal val clipsViewPagerImplGetViewAtIndexFingerprint = Fingerprint(
+internal object ClipsViewPagerImplGetViewAtIndexFingerprint : Fingerprint(
     strings = listOf("ClipsViewPagerImpl_getViewAtIndex")
 )
 
-internal val clipsSwipeRefreshLayoutOnInterceptTouchEventFingerprint  = Fingerprint (
+internal object ClipsSwipeRefreshLayoutOnInterceptTouchEventFingerprint : Fingerprint (
     parameters = listOf("Landroid/view/MotionEvent;"),
     definingClass = "Linstagram/features/clips/viewer/ui/ClipsSwipeRefreshLayout;"
 )