+++ patches/src/main/kotlin/app/morphe/patches/mapy/misc/spoof/Fingerprints.kt
@@ -1,7 +1,0 @@
-package app.morphe.patches.mapy.misc.spoof
-
-import app.morphe.patcher.fingerprint
-
-internal val securityCheckFingerprint = fingerprint {
-    custom { method, _ ->  method.name == "runSecurityChecks"}
-}